backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
pgvector
redis
celery
python-dotenv
pydantic-settings
httpx
requests
scapy
netifaces
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
google-generativeai
pydantic-ai
python-nmap
openai
mac-vendor-lookup
psutil
